fc_worker/requirements.txt
motor==3.6.0
ultralytics==8.3.67
Pillow==10.3.0
oss2==2.18.4